# book-network/src/main/java/com/hako/book/feedback/FeedbackResponse.java
package com.hako.book.feedback;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {

  private Double note;
  private String comment;
  private boolean ownFeedBack;

}